Cambios de formato en ABC *responde a confirmacion*
Se mejoró la modificación, y la respuesta al usuario

+++ public_html/procesarModificacion.php
@@ -8,13 +8,12 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
 
     // Acceder a los datos como un array asociativo
     $id = $_POST['id'];
-    $nombre = $_POST['nombre'];
-    $categoria = $_POST['categoria'];
-    $foto = $_FILES['fotoMod']; // Obtiene el archivo de imagen
-    $cantidad = $_POST['cantidad'];
-    $precio = $_POST['precio'];
-    $descuento = $_POST['descuento'];
-    $descripcion = $_POST['descripcion'];
+    $nombre = $_POST['NombreP'];
+    $categoria = $_POST['CategoriaP'];
+    $cantidad = $_POST['CantidadP'];
+    $precio = $_POST['PrecioP'];
+    $descuento = $_POST['DescuentoP'];
+    $descripcion = $_POST['DescripcionP'];
     $PrecioN = $precio - ($precio * $descuento / 100);
 
     // Consulta para obtener la ruta de la imagen existente
@@ -23,20 +22,30 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     $stmt->execute();
     $result = $stmt->get_result();
     $row = $result->fetch_assoc();
-    $ruta_imagen_existente = $row['imagen'];
+    $rutaPredefinida = "img/productos/";
+    $nombreImagen = $row['imagen'];
+    $ruta_imagen_existente = $rutaPredefinida . $nombreImagen;
 
-    // Borrar la imagen existente
-    if (file_exists($ruta_imagen_existente)) {
-        unlink($ruta_imagen_existente);
+    // Verificar si se ha enviado una nueva imagen
+    if (isset($_FILES['fotoMod']) && $_FILES['fotoMod']['error'] === UPLOAD_ERR_OK) {
+        $foto = $_FILES['fotoMod']; // Obtiene el archivo de imagen
+
+        // Borrar la imagen existente
+        if (file_exists($ruta_imagen_existente)) {
+            echo $ruta_imagen_existente;
+            unlink($ruta_imagen_existente);
+        }
+
+        // Mover la nueva imagen a la carpeta deseada
+        $ruta_destino = "img/productos/" . basename($foto['name']);
+        move_uploaded_file($foto['tmp_name'], $ruta_destino);
+
+        $nombreImagen = basename($foto['name']);
     }
-
-    // Mover la nueva imagen a la carpeta deseada
-    $ruta_destino = "img/productos/" . basename($foto['name']);
-    move_uploaded_file($foto['tmp_name'], $ruta_destino);
 
     // Actualizar la base de datos con la nueva información
     $statement = $conexion->prepare("UPDATE productos SET Nombre = ?, Categoria = ?, imagen = ?, Cantidad = ?, Precio = ?, Descuento = ?, Descripcion = ?, PrecioN = ? WHERE id = ?");
-    $statement->bind_param("sssiiisii", $nombre, $categoria, $ruta_destino, $cantidad, $precio, $descuento, $descripcion, $PrecioN, $id);
+    $statement->bind_param("sssiiisii", $nombre, $categoria, $nombreImagen, $cantidad, $precio, $descuento, $descripcion, $PrecioN, $id);
     $statement->execute();
 
     if ($statement->affected_rows == 1) {
